requirements.txt
anthropic
python-dotenv
duckduckgo-search
beautifulsoup4
httpx
html2text
requests
rich
